requirements.txt
groq
pandas
numpy
scikit-learn
matplotlib
seaborn
jupyter
notebook
streamlit
plotly
requests
python-dotenv
